session_log: carve bus_events.rs (pure move)
The event-bus-driven logging methods of SessionLog leave the big
inherent impl into a reopened impl block (inherent impls may split
across modules): session lifecycle/steer/approval/display/recording/
usage writers, turn artifacts (model responses, agent I/O, context
snapshots), and summary/interrupt markers. Writer-driven tests ride
along; read_last_event stays a mod.rs test fixture imported by both
this slice's and replay's tests.
